refactor(migrate): move the eighteen inline test modules to sibling files (#1918)
Sixth crate of the per-crate migration campaign, and the most uniform so
far: all eighteen blocks are a plain #[cfg(test)] mod tests at end of
file, top-level, with no pre-existing sibling to collide with, so every
one is a straight #[path] move.

The two mod.rs files get descriptive siblings — connectors_tests.rs and
wizard_tests.rs — rather than the tool's default mod_tests.rs, which
would have put two identically named files in the tree.

rustfmt reflows a few statements that fit on one line once dedented by
four columns; the sorted test-name list is byte-identical before and
after (263 lib tests, all passing).

The inline-tests baseline shrinks by the eighteen velesdb-migrate lines
(3310 lines paid), leaving 129 files.
