+++ crates/velesdb-migrate/src/wizard/wizard_tests.rs
@@ -1,0 +1,417 @@
+use super::*;
+use crate::config::{DistanceMetric, StorageMode};
+
+// ==================== SourceType Tests ====================
+
+#[test]
+fn test_source_type_all_returns_all_variants() {
+    let all = SourceType::all();
+    assert_eq!(all.len(), 10);
+    assert!(all.contains(&SourceType::Supabase));
+    assert!(all.contains(&SourceType::Qdrant));
+    assert!(all.contains(&SourceType::Pinecone));
+    assert!(all.contains(&SourceType::Weaviate));
+    assert!(all.contains(&SourceType::Milvus));
+    assert!(all.contains(&SourceType::ChromaDB));
+    assert!(all.contains(&SourceType::JsonFile));
+    assert!(all.contains(&SourceType::CsvFile));
+    assert!(all.contains(&SourceType::Elasticsearch));
+    assert!(all.contains(&SourceType::Redis));
+}
+
+#[test]
+fn test_source_type_display_names() {
+    assert_eq!(
+        SourceType::Supabase.display_name(),
+        "Supabase (PostgreSQL + pgvector)"
+    );
+    assert_eq!(SourceType::Qdrant.display_name(), "Qdrant");
+    assert_eq!(SourceType::Pinecone.display_name(), "Pinecone");
+    assert_eq!(SourceType::Weaviate.display_name(), "Weaviate");
+    assert_eq!(SourceType::Milvus.display_name(), "Milvus / Zilliz Cloud");
+    assert_eq!(SourceType::ChromaDB.display_name(), "ChromaDB");
+}
+
+#[test]
+fn test_source_type_short_names() {
+    assert_eq!(SourceType::Supabase.short_name(), "supabase");
+    assert_eq!(SourceType::Qdrant.short_name(), "qdrant");
+    assert_eq!(SourceType::Pinecone.short_name(), "pinecone");
+    assert_eq!(SourceType::Weaviate.short_name(), "weaviate");
+    assert_eq!(SourceType::Milvus.short_name(), "milvus");
+    assert_eq!(SourceType::ChromaDB.short_name(), "chromadb");
+}
+
+#[test]
+fn test_source_type_requires_api_key() {
+    // Sources that require an API key.
+    assert!(SourceType::Supabase.requires_api_key());
+    assert!(SourceType::Pinecone.requires_api_key());
+
+    // Sources that do not require an API key.
+    assert!(!SourceType::Qdrant.requires_api_key());
+    assert!(!SourceType::Weaviate.requires_api_key());
+    assert!(!SourceType::Milvus.requires_api_key());
+    assert!(!SourceType::ChromaDB.requires_api_key());
+}
+
+#[test]
+fn test_source_type_optional_api_key() {
+    // Sources with an optional API key.
+    assert!(SourceType::Qdrant.optional_api_key());
+    assert!(SourceType::Weaviate.optional_api_key());
+    assert!(SourceType::Milvus.optional_api_key());
+
+    // Sources without an optional API key.
+    assert!(!SourceType::Supabase.optional_api_key());
+    assert!(!SourceType::Pinecone.optional_api_key());
+    assert!(!SourceType::ChromaDB.optional_api_key());
+}
+
+// ==================== WizardConfig Tests ====================
+
+#[test]
+fn test_wizard_config_creation() {
+    // Arrange & Act
+    let config = WizardConfig {
+        source_type: SourceType::Qdrant,
+        url: "http://localhost:6333".to_string(),
+        api_key: None,
+        collection: "test_collection".to_string(),
+        dest_path: "./velesdb_data".to_string(),
+        use_sq8: false,
+    };
+
+    // Assert
+    assert_eq!(config.source_type, SourceType::Qdrant);
+    assert_eq!(config.url, "http://localhost:6333");
+    assert!(config.api_key.is_none());
+    assert_eq!(config.collection, "test_collection");
+    assert_eq!(config.dest_path, "./velesdb_data");
+    assert!(!config.use_sq8);
+}
+
+#[test]
+fn test_wizard_config_with_api_key() {
+    // Arrange & Act
+    let config = WizardConfig {
+        source_type: SourceType::Supabase,
+        url: "https://xyz.supabase.co".to_string(),
+        api_key: Some("secret-key".to_string()),
+        collection: "documents".to_string(),
+        dest_path: "./data".to_string(),
+        use_sq8: true,
+    };
+
+    // Assert
+    assert_eq!(config.source_type, SourceType::Supabase);
+    assert_eq!(config.api_key, Some("secret-key".to_string()));
+    assert!(config.use_sq8);
+}
+
+// ==================== Wizard Tests ====================
+
+#[test]
+fn test_wizard_new() {
+    // Arrange & Act
+    let wizard = Wizard::new();
+
+    // Assert - wizard should be created without panic
+    // We can't easily test internal state, but creation should work
+    drop(wizard);
+}
+
+#[test]
+fn test_wizard_default() {
+    // Arrange & Act
+    let wizard = Wizard::default();
+
+    // Assert
+    drop(wizard);
+}
+
+// ==================== Build Source Config Tests ====================
+
+#[test]
+fn test_build_source_config_qdrant() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::Qdrant,
+        url: "http://localhost:6333".to_string(),
+        api_key: Some("test-key".to_string()),
+        collection: "vectors".to_string(),
+        dest_path: "./data".to_string(),
+        use_sq8: false,
+    };
+
+    // Act
+    let result = wizard.build_source_config(&config);
+
+    // Assert
+    assert!(result.is_ok());
+    let source = result.unwrap();
+    match source {
+        SourceConfig::Qdrant(cfg) => {
+            assert_eq!(cfg.url, "http://localhost:6333");
+            assert_eq!(cfg.collection, "vectors");
+            assert_eq!(cfg.api_key, Some("test-key".to_string()));
+        }
+        _ => panic!("Expected Qdrant config"),
+    }
+}
+
+#[test]
+fn test_build_source_config_supabase() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::Supabase,
+        url: "https://xyz.supabase.co".to_string(),
+        api_key: Some("service-key".to_string()),
+        collection: "documents".to_string(),
+        dest_path: "./data".to_string(),
+        use_sq8: false,
+    };
+
+    // Act
+    let result = wizard.build_source_config(&config);
+
+    // Assert
+    assert!(result.is_ok());
+    let source = result.unwrap();
+    match source {
+        SourceConfig::Supabase(cfg) => {
+            assert_eq!(cfg.url, "https://xyz.supabase.co");
+            assert_eq!(cfg.table, "documents");
+            assert_eq!(cfg.api_key, "service-key");
+        }
+        _ => panic!("Expected Supabase config"),
+    }
+}
+
+#[test]
+fn test_build_source_config_chromadb() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::ChromaDB,
+        url: "http://localhost:8000".to_string(),
+        api_key: None,
+        collection: "embeddings".to_string(),
+        dest_path: "./data".to_string(),
+        use_sq8: false,
+    };
+
+    // Act
+    let result = wizard.build_source_config(&config);
+
+    // Assert
+    assert!(result.is_ok());
+    let source = result.unwrap();
+    match source {
+        SourceConfig::ChromaDB(cfg) => {
+            assert_eq!(cfg.url, "http://localhost:8000");
+            assert_eq!(cfg.collection, "embeddings");
+        }
+        _ => panic!("Expected ChromaDB config"),
+    }
+}
+
+#[test]
+fn test_build_source_config_pinecone() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::Pinecone,
+        url: "https://index.pinecone.io".to_string(),
+        api_key: Some("pinecone-key".to_string()),
+        collection: "my-index".to_string(),
+        dest_path: "./data".to_string(),
+        use_sq8: false,
+    };
+
+    // Act
+    let result = wizard.build_source_config(&config);
+
+    // Assert
+    assert!(result.is_ok());
+    let source = result.unwrap();
+    match source {
+        SourceConfig::Pinecone(cfg) => {
+            assert_eq!(cfg.api_key, "pinecone-key");
+            assert_eq!(cfg.index, "my-index");
+        }
+        _ => panic!("Expected Pinecone config"),
+    }
+}
+
+#[test]
+fn test_build_source_config_weaviate() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::Weaviate,
+        url: "http://localhost:8080".to_string(),
+        api_key: None,
+        collection: "Document".to_string(),
+        dest_path: "./data".to_string(),
+        use_sq8: false,
+    };
+
+    // Act
+    let result = wizard.build_source_config(&config);
+
+    // Assert
+    assert!(result.is_ok());
+    let source = result.unwrap();
+    match source {
+        SourceConfig::Weaviate(cfg) => {
+            assert_eq!(cfg.url, "http://localhost:8080");
+            assert_eq!(cfg.class_name, "Document");
+        }
+        _ => panic!("Expected Weaviate config"),
+    }
+}
+
+#[test]
+fn test_build_source_config_milvus() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::Milvus,
+        url: "http://localhost:19530".to_string(),
+        api_key: None,
+        collection: "vectors".to_string(),
+        dest_path: "./data".to_string(),
+        use_sq8: false,
+    };
+
+    // Act
+    let result = wizard.build_source_config(&config);
+
+    // Assert
+    assert!(result.is_ok());
+    let source = result.unwrap();
+    match source {
+        SourceConfig::Milvus(cfg) => {
+            assert_eq!(cfg.url, "http://localhost:19530");
+            assert_eq!(cfg.collection, "vectors");
+        }
+        _ => panic!("Expected Milvus config"),
+    }
+}
+
+// ==================== Build Migration Config Tests ====================
+
+#[test]
+fn test_build_migration_config_full_storage() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::Qdrant,
+        url: "http://localhost:6333".to_string(),
+        api_key: None,
+        collection: "test".to_string(),
+        dest_path: "./velesdb_data".to_string(),
+        use_sq8: false,
+    };
+    let schema = crate::connectors::SourceSchema {
+        source_type: "qdrant".to_string(),
+        collection: "test".to_string(),
+        dimension: 768,
+        total_count: Some(1000),
+        fields: vec![],
+        vector_column: None,
+        id_column: None,
+        metric: None,
+    };
+
+    // Act
+    let result = wizard.build_migration_config(&config, &schema);
+
+    // Assert
+    assert!(result.is_ok());
+    let migration = result.unwrap();
+    assert_eq!(migration.destination.dimension, 768);
+    assert_eq!(migration.destination.collection, "test");
+    assert!(matches!(
+        migration.destination.storage_mode,
+        StorageMode::Full
+    ));
+    assert!(matches!(
+        migration.destination.metric,
+        DistanceMetric::Cosine
+    ));
+}
+
+#[test]
+fn test_build_migration_config_sq8_storage() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::Qdrant,
+        url: "http://localhost:6333".to_string(),
+        api_key: None,
+        collection: "test".to_string(),
+        dest_path: "./velesdb_data".to_string(),
+        use_sq8: true, // SQ8 enabled
+    };
+    let schema = crate::connectors::SourceSchema {
+        source_type: "qdrant".to_string(),
+        collection: "test".to_string(),
+        dimension: 1536,
+        total_count: Some(5000),
+        fields: vec![],
+        vector_column: None,
+        id_column: None,
+        metric: None,
+    };
+
+    // Act
+    let result = wizard.build_migration_config(&config, &schema);
+
+    // Assert
+    assert!(result.is_ok());
+    let migration = result.unwrap();
+    assert_eq!(migration.destination.dimension, 1536);
+    assert!(matches!(
+        migration.destination.storage_mode,
+        StorageMode::SQ8
+    ));
+}
+
+#[test]
+fn test_build_migration_config_options() {
+    // Arrange
+    let wizard = Wizard::new();
+    let config = WizardConfig {
+        source_type: SourceType::ChromaDB,
+        url: "http://localhost:8000".to_string(),
+        api_key: None,
+        collection: "embeddings".to_string(),
+        dest_path: "./data".to_string(),
+        use_sq8: false,
+    };
+    let schema = crate::connectors::SourceSchema {
+        source_type: "chromadb".to_string(),
+        collection: "embeddings".to_string(),
+        dimension: 384,
+        total_count: None,
+        fields: vec![],
+        vector_column: None,
+        id_column: None,
+        metric: None,
+    };
+
+    // Act
+    let result = wizard.build_migration_config(&config, &schema);
+
+    // Assert
+    assert!(result.is_ok());
+    let migration = result.unwrap();
+    assert_eq!(migration.options.batch_size, 1000);
+    assert_eq!(migration.options.workers, 4);
+    assert!(!migration.options.dry_run);
+    assert!(migration.options.checkpoint_enabled);
+}
