+++ scripts/inline-tests-baseline.txt
@@ -79,24 +79,6 @@ crates/velesdb-core/src/velesql/error.rs	67
 crates/velesdb-core/src/velesql/parser/helpers.rs	144
 crates/velesdb-core/src/velesql/parser/hints.rs	49
 crates/velesdb-core/src/wire/vrb1.rs	119
-crates/velesdb-migrate/src/config.rs	202
-crates/velesdb-migrate/src/connectors/chromadb.rs	47
-crates/velesdb-migrate/src/connectors/common.rs	271
-crates/velesdb-migrate/src/connectors/milvus.rs	271
-crates/velesdb-migrate/src/connectors/mod.rs	42
-crates/velesdb-migrate/src/connectors/pinecone.rs	185
-crates/velesdb-migrate/src/connectors/qdrant.rs	152
-crates/velesdb-migrate/src/connectors/relation_detector.rs	58
-crates/velesdb-migrate/src/connectors/supabase.rs	89
-crates/velesdb-migrate/src/connectors/weaviate.rs	76
-crates/velesdb-migrate/src/error.rs	17
-crates/velesdb-migrate/src/pipeline.rs	231
-crates/velesdb-migrate/src/pipeline_graph.rs	219
-crates/velesdb-migrate/src/retry.rs	308
-crates/velesdb-migrate/src/source_config_builder.rs	291
-crates/velesdb-migrate/src/templates.rs	332
-crates/velesdb-migrate/src/transform.rs	99
-crates/velesdb-migrate/src/wizard/mod.rs	420
 crates/velesdb-server/src/auth.rs	115
 crates/velesdb-server/src/config.rs	559
 crates/velesdb-server/src/handlers/admin.rs	88
